v8.3.0

+++ Core/src/main/java/su/nightexpress/dungeons/dungeon/script/condition/impl/AliveMobAmountCondition.java
@@ -25,7 +25,7 @@ public class AliveMobAmountCondition extends MobAmountCondition {
     }
 
     @Override
-    protected int getDungeonValue(@NotNull DungeonInstance dungeon) {
+    protected double getDungeonValue(@NotNull DungeonInstance dungeon) {
         return dungeon.countMobs(mob -> mob.isMob(this.identifier));
     }
 }
